backend_service/requirements.txt
flask>=3.0.0
psycopg2-binary>= 2.9.9
python-dotenv>=1.0.0
Flask-SQLAlchemy>=3.1.1
flask-bcrypt>=1.0.1
flask-session>=0.5.0
redis>=5.0.1
flask-cors>=4.0.0